+++ app/src/main/java/com/drawiin/mymovielist/common/movie/domain/usecase/AddMovieToWatchListUseCase.kt
@@ -1,0 +1,7 @@
+package com.drawiin.mymovielist.common.movie.domain.usecase
+
+import com.drawiin.mymovielist.common.movie.domain.model.MovieModel
+
+fun interface AddMovieToWatchListUseCase {
+    suspend operator fun invoke(movie: MovieModel): Result<Unit>
+}