+++ app/src/main/java/com/drawiin/mymovielist/core/navigation/AppRoutes.kt
@@ -14,5 +14,7 @@ sealed interface AppRoutes {
         val movieId: Int
     ) : AppRoutes
 
+    @Serializable
+    object WatchListRoute : AppRoutes
 }