+++ app/src/main/java/com/drawiin/mymovielist/feature/list/data/datasource/remote/MovieListRemoteDataSource.kt
@@ -1,0 +1,7 @@
+package com.drawiin.mymovielist.feature.list.data.datasource.remote
+
+import com.drawiin.mymovielist.feature.list.data.dto.MovieListResponseDTO
+
+interface MovieListRemoteDataSource {
+    suspend fun getPopularMovies(page: Int): Result<MovieListResponseDTO>
+}