+++ app/src/main/java/com/drawiin/mymovielist/feature/list/data/datasource/remote/MovieListRemoteDataSourceImpl.kt
@@ -1,4 +1,4 @@
-package com.drawiin.mymovielist.feature.list.data.datasource
+package com.drawiin.mymovielist.feature.list.data.datasource.remote
 
 import com.drawiin.mymovielist.core.network.client.MyMovieListNetworkClient
 import com.drawiin.mymovielist.feature.list.data.dto.MovieListResponseDTO
@@ -7,9 +7,9 @@ import javax.inject.Inject
 class MovieListRemoteDataSourceImpl @Inject constructor(
     val client: MyMovieListNetworkClient
 ): MovieListRemoteDataSource {
-    override suspend fun getPopularMovies(): Result<MovieListResponseDTO> {
+    override suspend fun getPopularMovies(page: Int): Result<MovieListResponseDTO> {
         return client.get(
-            "popular?language=en-US&page=1",
+            "popular?language=en-US&page=${page}",
             MovieListResponseDTO::class
         )
     }