+++ app/src/main/java/com/drawiin/mymovielist/feature/list/data/repository/MovieListRepositoryImpl.kt
@@ -1,18 +1,16 @@
 package com.drawiin.mymovielist.feature.list.data.repository
 
-import com.drawiin.mymovielist.common.movie.domain.model.MoviePreviewModel
-import com.drawiin.mymovielist.feature.list.data.datasource.MovieListRemoteDataSource
+import com.drawiin.mymovielist.feature.list.data.datasource.remote.MovieListRemoteDataSource
 import com.drawiin.mymovielist.feature.list.data.dto.toDomainModel
+import com.drawiin.mymovielist.feature.list.domain.model.MovieListPageModel
 import javax.inject.Inject
 
 class MovieListRepositoryImpl @Inject constructor(
     private val remoteDataSource: MovieListRemoteDataSource
 ) : MovieListRepository {
-    override suspend fun getMovies(): Result<List<MoviePreviewModel>> {
-        return remoteDataSource.getPopularMovies().map {
-            it.results.map { movie ->
-                movie.toDomainModel()
-            }
+    override suspend fun getMovies(page: Int): Result<MovieListPageModel> {
+        return remoteDataSource.getPopularMovies(page).map {
+            it.toDomainModel()
         }
     }
 }